Ontologies

The rols package provides an interface to PRIDE's Ontology Lookup Servive (OLS) and can be used to query one or multiple ontologies, stored as `Ontology` and `Ontologies` instances, and containing various information as provided by OLS.

# S4 method for missing
Ontologies(object)

# S4 method for character
Ontology(object)

# S4 method for Ontology
Ontology(object)

# S4 method for Ontology
show(object)

# S4 method for Ontologies
show(object)

# S4 method for character
olsVersion(object)

# S4 method for Ontology
olsVersion(object)

# S4 method for Ontologies
olsVersion(object)

# S4 method for character
olsLoaded(object)

# S4 method for Ontology
olsLoaded(object)

# S4 method for Ontologies
olsLoaded(object)

# S4 method for Ontology
olsLinks(object)

# S4 method for Ontology
olsConfig(object)

# S4 method for character
olsUpdated(object)

# S4 method for Ontology
olsUpdated(object)

# S4 method for Ontologies
olsUpdated(object)

# S4 method for character
olsPrefix(object)

# S4 method for Ontology
olsPrefix(object)

# S4 method for Ontologies
olsPrefix(object)

# S4 method for character
olsDesc(object)

# S4 method for Ontology
olsDesc(object)

# S4 method for Ontologies
olsDesc(object)

# S4 method for character
olsTitle(object)

# S4 method for Ontology
olsTitle(object)

# S4 method for Ontologies
olsTitle(object)

# S4 method for character
olsStatus(object)

# S4 method for Ontology
olsStatus(object)

# S4 method for Ontologies
olsStatus(object)

# S4 method for character
olsNamespace(object)

# S4 method for Ontology
olsNamespace(object)

# S4 method for Ontologies
olsNamespace(object)

# S4 method for character
ontologyUrl(object)

# S4 method for Ontology
ontologyUrl(object)

# S4 method for Ontologies
lapply(X, FUN, ...)

# S4 method for Ontologies
[(x, i, j = "missing", drop = "missing")

# S4 method for Ontologies
[[(x, i, j = "missing", drop = "missing")

# S4 method for Ontologies
length(x)

Arguments

object

an instance of class `Ontologies` or `Ontology`. For some functions, a ontology identifier is applicable.

X

`Ontologies` object.

FUN

a `function` to be applied to each `Ontology` element of `X`.

...

additional arguments passed to `FUN`.

x

an `Ontologies` object.

i

index of elecements to subset.

j

ignored.

drop

ignored.

Details

Ontologies are referred to by their namespace, which is lower case: the Gene Onology is "go", the Mass spectrometry ontology is "ms", etc. The ontologies also have prefixes, which are upper case: the Gene Onology prefix "GO", the Mass spectrometry ontology prefix "MS". One exception to this rule is the Drosophila Phenotype Ontology, whose namespace and prefix are "dpo" and "FBcv" respectively (there might be more). This is particularly confusing as the FlyBase Controlled Vocabulary has "fbcv" and "FBcv" as namespace and prefix respectively.

When using a character to initialise an ontology or query a term, "fbcv" (this is case insensitive) will refer to the the FlyBase Controlled Vocabulary. The the Drosophila Phenotype Ontology will have to be referred as "dpo" (also case insensitive).

Constructors

Objects can be created in multiple ways. The [Ontologies()] function will initialise all available ontolgies as an `Ontologies` object, while a call to [Ontology()] with an ontology namespace or prefix as argument will initialise the ontology of interest as an `Ontology` instance.

`Ontolgies` instances can be subset with `[` and `[[` (using their namespace, see Details) and iterated over with `lapply`. `Ontolgies` can be converted into a simple `data.frame` containing the ontology prefixes, namespaces and titles using `as(., "data.frame")`. `Ontologies` can also be coerced to lists of `Ontology` ojects with `as(., "list")`.

Accessors

- `olsDesc(object = "Ontology")` returns the description of an ontology. Also works for `Ontologies` objects and a `character` describing an ontology namespace or prefix (see Details).

- `olsPrefix(object = "Ontology")` retruns the prefix of an ontology. Also works for `Ontologies` objects and a `character` describing an ontology namespace or prefix (see Details).

- `olsVersion(object = "Ontology")` returns the version of the ontology. Also works with an a `character` defining an ontology namespace or prefix (see Details) or an object of class `Ontologies`, in which case it returns a list of versions.

- `olsLoaded(object = "Ontology")` returns the loading date of the ontology. Also works with a `character` containing the ontology namespace or prefix (see Details) or an object of class `Ontologies`.

- `olsUpdated(object = "Ontology")` returns the update date of the ontology. Also works with a `character` containing the ontology namespace or prefix (see Details) or an object of class `Ontologies`.

- `olsStatus(object = "Ontology")` returns the status of the ontology. Also works with a `character` containing the ontology namespace or prefix (see Details) or an object of class `Ontologies`.

- `olsTitle(object = "Ontology")` returns the title of an ontology. Also works with a `character` containing the ontology namespace or prefix (see Details) or an object of class `Ontologies`.

- `olsNamespace(object = "Ontology")` returns the namespace of an ontology. Also works with a `character` containing the ontology namespace or prefix (see Details) or an object of class `Ontologies`.

- `olsLinks(object = "Ontology")` returns a named `character` with hyperlink to the ontology itself, and other associated concepts such as its terms.

- `olsConfig(object = "Ontology")` returns a list of additional unstructured, partly redundant information about the ontology.

- `ontologyUrl(object = "Ontology") return the hyperlink to the ontology itself. It can also be used with a `character` defining the namespace or prefix of an ontology, in which case it is created from the base OLS API URL.

Ontology terms

Once an ontology has been created an an `Ontology` instance, all its terms can be requested using the `Terms()` constructor. See [Terms()] for details.

Author

Laurent Gatto

Examples

#############################
## All ontologies
(onts <- Ontologies())
#> Object of class 'Ontologies' with 250 entries
#>    ADO, AGRO ... CCF, CPONT 

#############################
## Alzheimer's Disease Ontology (ADO)
## 1. From the ontologies object
(ado1 <- onts[['ado']])
#> Ontology: Alzheimer's Disease Ontology (ADO) (ado)  
#>   Alzheimer's Disease Ontology is a knowledge-based ontology that
#>   encompasses varieties of concepts related to Alzheimer'S Disease,
#>   foundamentally structured by upper level Basic Formal Ontology(BFO).
#>   This Ontology is enriched by the interrelational entities that
#>   demonstrate the nextwork of the understanding on Alzheimer's disease
#>   and can be readily applied for text mining.
#>    Loaded: 2024-02-23 Updated: 2024-02-23 Version: 2.0.1 
#>    1963 terms  186 properties  2 individuals
## 2. Create from its namespace
(ado2 <- Ontology('ado')) ## also works with ADO
#> Ontology: Alzheimer's Disease Ontology (ADO) (ado)  
#>   Alzheimer's Disease Ontology is a knowledge-based ontology that
#>   encompasses varieties of concepts related to Alzheimer'S Disease,
#>   foundamentally structured by upper level Basic Formal Ontology(BFO).
#>   This Ontology is enriched by the interrelational entities that
#>   demonstrate the nextwork of the understanding on Alzheimer's disease
#>   and can be readily applied for text mining.
#>    Loaded: 2024-02-23 Updated: 2024-02-23 Version: 2.0.1 
#>    1963 terms  186 properties  2 individuals

all.equal(ado1, ado2)
#> [1] TRUE

olsVersion(ado1)
#> [1] "2.0.1"
olsPrefix(ado1)
#> [1] "ADO"
olsNamespace(ado1)
#> [1] "ado"
olsTitle(ado1)
#> [1] "Alzheimer's Disease Ontology (ADO)"
olsDesc(ado1)
#> [1] "Alzheimer's Disease Ontology is a knowledge-based ontology that encompasses varieties of concepts related to Alzheimer'S Disease, foundamentally structured by upper level Basic Formal Ontology(BFO). This Ontology is enriched by the interrelational entities that demonstrate the nextwork of the understanding on Alzheimer's disease and can be readily applied for text mining."
olsLinks(ado1)
#>                                                        self 
#>     "https://www.ebi.ac.uk/ols4/api/ontologies/ado?lang=en" 
#>                                                       terms 
#>       "https://www.ebi.ac.uk/ols4/api/ontologies/ado/terms" 
#>                                                  properties 
#>  "https://www.ebi.ac.uk/ols4/api/ontologies/ado/properties" 
#>                                                 individuals 
#> "https://www.ebi.ac.uk/ols4/api/ontologies/ado/individuals" 
str(olsConfig(ado1))
#> List of 24
#>  $ id                    : chr "ado"
#>  $ versionIri            : chr "http://purl.obolibrary.org/obo/ado/releases/2023-09-20/ado.owl"
#>  $ namespace             : chr "ado"
#>  $ preferredPrefix       : chr "ADO"
#>  $ title                 : chr "Alzheimer's Disease Ontology (ADO)"
#>  $ description           : chr "Alzheimer's Disease Ontology is a knowledge-based ontology that encompasses varieties of concepts related to Al"| __truncated__
#>  $ homepage              : chr "https://github.com/Fraunhofer-SCAI-Applied-Semantics/ADO"
#>  $ version               : chr "2.0.1"
#>  $ mailingList           : NULL
#>  $ tracker               : chr "https://github.com/Fraunhofer-SCAI-Applied-Semantics/ADO/issues"
#>  $ logo                  : NULL
#>  $ creators              : list()
#>  $ annotations           : NULL
#>  $ fileLocation          : chr "http://purl.obolibrary.org/obo/ado.owl"
#>  $ oboSlims              : logi FALSE
#>  $ labelProperty         : chr "http://www.w3.org/2000/01/rdf-schema#label"
#>  $ definitionProperties  : list()
#>  $ synonymProperties     : list()
#>  $ hierarchicalProperties: list()
#>  $ baseUris              : list()
#>  $ hiddenProperties      : list()
#>  $ preferredRootTerms    : list()
#>  $ isSkos                : logi FALSE
#>  $ allowDownload         : logi FALSE