Term Properties

Properties (relationships) between terms can be queries for complete [Ontology()] objects and [Term()]/[Terms()] instances, and the results are stored as objects of class `Property` or `Properties`.

# S4 method for Ontology
Properties(object)

# S4 method for character
Properties(object)

# S4 method for Term
Properties(object)

# S4 method for Terms
Properties(object)

# S4 method for Property
show(object)

# S4 method for Properties
show(object)

# S4 method for Properties
length(x)

Arguments

object

First input object.

x

A `Properties` object.

Examples

## Term properties
trm <- Term("uberon", "UBERON:0002107")
trm
#> A Term from the UBERON ontology: UBERON:0002107 
#>  Label: liver
#>   An exocrine gland which secretes bile and functions in metabolism of
#>   protein and carbohydrate and fat, synthesizes substances involved in
#>   the clotting of the blood, synthesizes vitamin A, detoxifies poisonous
#>   substances, stores glycogen, and breaks down worn-out erythrocytes[GO].
Properties(trm)
#> Object of class 'Properties' with 158 entries
#>  From the UBERON ontology
#>   hepatobiliary system, exocrine system ... liver serosa, liver subserosa 

## Ontology properties
Properties('ado')
#> ⠙ Iterating 9 done (4.4/s) | 2.1s
#> ⠙ Iterating 10 done (4.5/s) | 2.2s
#> Object of class 'Properties' with 186 entries
#>  From the ADO ontology
#>   treeView, readOnly ... SNOMED CT Identifier, Library of Congress Subject Headings