Class "Properties"

Propteries (relationships) between terms can be queries for complete Ontology objects and Term/Terms instances, and the results are stored as objects of class Property or Properties.

Objects from the Class

Objects can be created by calls to properties, as described below.

Slots

See the Term and Terms classes.

Extends

Class "Terms", directly.

Methods and functions

properties

signature(object = "Ontology", pagesize = 200): ... Also works with a character with the ontology namespace. See Ontology for details.

properties

signature(object = "Term"): retrieves the properties of term object and returns a Properties object. Returns NULL when no properties are available.

proterties

signature(object = "Terms", ...): retrieves the properties of each term of object and returns a list of Properties (or NULL) items.

show

signature(object = "Properties"): shows a textual summary of the object.

length

signature(object = "Properties"): returns the number of properties in object.

Examples

trm <- term("uberon", "UBERON:0002107")
trm

#> A Term from the UBERON ontology: UBERON:0002107 
#>  Label: liver
#>   An exocrine gland which secretes bile and functions in metabolism of
#>   protein and carbohydrate and fat, synthesizes substances involved in
#>   the clotting of the blood, synthesizes vitamin A, detoxifies poisonous
#>   substances, stores glycogen, and breaks down worn-out erythrocytes[GO].

properties(trm)

#> Object of class 'Properties' with 116 entries
#>  From the UBERON ontology
#>   digestive system gland, abdomen element ... hepatobiliary system, liver lobule 

trm2 <- term("GO", "GO:0005326")
trm2

#> A Term from the GO ontology: GO:0005326 
#>  Label: neurotransmitter transporter activity
#>   Enables the directed movement of a neurotransmitter into, out of or
#>   within a cell, or between cells. Neurotransmitters are any chemical
#>   substance that is capable of transmitting (or inhibiting the
#>   transmission of) a nerve impulse from a neuron to another cell.

properties(trm2)

#> Object of class 'Properties' with 18 entries
#>  From the GO ontology
#>   transporter activity, neurotransmitter transport ... norepinephrine:sodium symporter activity, serotonin:sodium symporter activity