Ontology Terms

The `Term` class describes an ontology term. A set of terms are instantiated as a `Terms` class.

# S4 method for character
Terms(object, pagesize = 1000, obsolete = NULL)

# S4 method for Ontology
Terms(object, pagesize = 1000, obsolete = NULL)

# S4 method for character
Term(object, id)

# S4 method for Ontology
Term(object, id)

children(object)

parents(object)

ancestors(object)

descendants(object)

# S4 method for Term
show(object)

# S4 method for Terms
show(object)

# S4 method for Term
termSynonym(object)

# S4 method for Terms
termSynonym(object)

# S4 method for Term
isObsolete(object)

# S4 method for Terms
isObsolete(object)

# S4 method for Term
isRoot(object)

# S4 method for Terms
isRoot(object)

# S4 method for Term
termLabel(object)

# S4 method for Terms
termLabel(object)

# S4 method for Term
termId(object)

# S4 method for Terms
termId(object)

# S4 method for Term
termLinks(object)

# S4 method for Term
termPrefix(object)

# S4 method for Terms
termPrefix(object)

# S4 method for Term
termDesc(object)

# S4 method for Terms
termDesc(object)

# S4 method for Term
termOntology(object)

# S4 method for Terms
termOntology(object)

# S4 method for Term
termNamespace(object)

# S4 method for Terms
termNamespace(object)

# S4 method for Terms
length(x)

# S4 method for Terms
unique(x)

# S4 method for Terms
[(x, i, j = "missing", drop = "missing")

# S4 method for Terms
[[(x, i, j = "missing", drop = "missing")

# S4 method for Terms
lapply(X, FUN, ...)

as.Term.data.frame(x)

as.Terms.data.frame(x)

Arguments

object

generally an instance of class `Terms` or `Term`. In some cases, an ontology identifier is applicable.

pagesize

`numeric(1)`, converted to an integer, defining the response page size. Default is 1000.

obsolete

`NULL` or `logical(1)` defining whether obsolete terms (`TRUE`), current terms (`FALSE`) or all (`NULL`, default) should be returned.

id

`character(1)` with the term's identifier.

x

a `Terms` object.

i

index of elecements to subset.

j

ignored.

drop

ignored.

X

`Terms` object.

FUN

a `function` to be applied to each `Term` element of `X`.

...

additional arguments passed to `FUN`.

Contructors

Objects can be created using the `Term()` and `Terms()` constructers. The latter is used with an object of class `Ontology` or a `character` describing a valid ontology prefix to download and instantiate all terms of an ontology of interest. The former takes an `Ontology` object (or an ontology prefix) and a term identifer to instantiate that specific term.

For any given `Term` object, the `children`, `parents`, `ancestors` and `descendants` terms can be generated with the `children()`, `parents()`, `ancestor()` and `descendants()` function. `Terms` instances can be subset with `[` and `[[` and iterated over with `lapply`.

Accessors

- `isObsolete(object = "Term")` returns a `TRUE` if the term is obsolete, `FALSE` otherwise. Also works on `Terms` instances.

- `isRoot(object = "Term")` returns a `TRUE` if the term is a root term, `FALSE` otherwise. Also works on `Terms` instances.

- `termDesc(object = "Term")` returns a `character` with the term's description. Also works on `Terms` instances.

- `termId(object = "Term")` returns a `character` with the term's identifier. Also works on `Terms` instances.

- `termLabel(object = "Term")` returns a `character` with the term's label. Also works on `Terms` instances.

- `termNamespace(object = "Term")` returns a `character` with the term's namespace. Also works on `Terms` instances.

- `termOntology(object = "Term")` returns a `character` with the term's ontology (where it was retrieved from). Also works on `Terms` instances.

- `termPrefix(object = "Term")` returns a `character` with the term's (ontology) prefix (where it was retrieved from). Also works on `Terms` instances.

- `termSynonym(object = "Term")` returns a `character` with the term's synpnym(s). Also works on `Terms` instances.

- `termLinks(object = "Term")` returns a named `character` with hyperlink to/from the term.

Related terms

- `children(object = "Term")` returns a new `Terms` instance with the `object`'s children or `NULL` if there are no children.

- `parents(object = "Term")` returns a new `Terms` instance with the `object`'s parents or `NULL` if there are no parents.

- `ancestors(object = "Term")` returns a new `Terms` instance with the `object`'s ancestors or `NULL` if there are no ancestors.

- `descendants(object = "Term")` returns a new `Terms` instance with the `object`'s descendants or `NULL` if there are no descendants.

Coercion

- `as(x, "data.fram")` coerces a `Term` or `Terms` instance into a `data.frame` of length 1 (for the former) or length `length(x)` for the latter. The result will contain the following columns: id, label, description of the term(s), their ontology, whether they are obsolete, have children or are root node, the first synonym only, their iri and whether they are defining the ontology. Any missing value will be reported as `NA`.

Author

Laurent Gatto

Examples

## Alzheimer's Disease Ontology (ADO)
(adoterms <- Terms('ado'))
#> ⠙ Iterating 1 done (0.09/s) | 11.1s
#> ⠹ Iterating 2 done (0.11/s) | 18.5s
#> Object of class 'Terms' with 1963 entries
#>  From the ADO ontology
#>   UBERON:0004120, UBERON:0004121 ... UBERON:0003947, UBERON:0004086 

## Focus on squamous epithelium
(trm <- adoterms[["UBERON:0006914"]])
#> A Term from the ADO ontology: UBERON:0006914 
#>  Label: squamous epithelium
#>   An epithelium characterised by its most superficial layer consisting of
#>   squamous epithelial cells.

## Accessors
termLabel(trm)
#> [1] "squamous epithelium"
head(termLabel(adoterms))
#>                UBERON:0004120                UBERON:0004121 
#>  "mesoderm-derived structure"  "ectoderm-derived structure" 
#>                UBERON:0004923                UBERON:0005162 
#>       "organ component layer"   "multi cell part structure" 
#>                UBERON:0005174                UBERON:0005358 
#>       "dorsal region element" "ventricle of nervous system" 
termId(trm)
#> [1] "UBERON:0006914"
termDesc(trm)
#> [1] "An epithelium characterised by its most superficial layer consisting of squamous epithelial cells."
termOntology(trm)
#> [1] "ado"
termNamespace(trm)
#> [1] "uberon"
termSynonym(trm) ## none
#> NULL

## Related terms
children(trm)
#> Object of class 'Terms' with 1 entries
#>  From the ADO ontology
#> UBERON:0000487 
descendants(trm) ## includes child
#> Object of class 'Terms' with 2 entries
#>  From the ADO ontology
#> UBERON:0000487, UBERON:0001986 

parents(trm)
#> Object of class 'Terms' with 1 entries
#>  From the ADO ontology
#> UBERON:0000483 
ancestors(trm) ## includes parent
#> Object of class 'Terms' with 10 entries
#>  From the ADO ontology
#>   UBERON:0000483, UBERON:0000479 ... BFO:0000001, UBERON:0001062 

## A single term from an ontology
Term("ado", "ADO:0000090")
#> A Term from the ADO ontology: ADO:0000090 
#>  Label: total tau
#>   A protein that is regarded as a general marker of neurodegeneration for
#>   evaluation in future population-based studies and represents the total
#>   tau concentration in a given sample.