Class "OlsSearch"

Searching the OLS is done using the OlsSearch data structure.

Objects from the Class

Objects can be created with the constructor function OlsSearch.

Slots

q:

Object of class "character" ~~

ontology:

Object of class "character" ~~

type:

Object of class "character" ~~

slim:

Object of class "character" ~~

fieldList:

Object of class "character" ~~

queryFields:

Object of class "character" ~~

exact:

Object of class "logical" ~~

groupField:

Object of class "logical" ~~

obsoletes:

Object of class "logical" ~~

local:

Object of class "character" ~~

childrenOf:

Object of class "character" ~~

rows:

Object of class "integer" ~~

start:

Object of class "integer" ~~

url:

Object of class "character" ~~

numFound:

Object of class "integer" ~~

response:

Object of class "data.frame" ~~

Methods and functions

coerce

signature(from = "OlsSearch", to = "data.frame"): ...

coerce

signature(from = "OlsSearch", to = "Terms"): ...

show

signature(object = "OlsSearch"): ...

olsRows

signature(object = "OlsSearch"): ... The value can be updated with the olsRows replacement method. To request all responses, use allRows.

Author

Laurent Gatto <lg390@cam.ac.uk>

Examples

OlsSearch(q = "trans-golgi")

#> Object of class 'OlsSearch':
#>   query: trans-golgi 
#>   requested: 20 (out of 12994)
#>   response(s): 0 

OlsSearch(q = "cell")

#> Object of class 'OlsSearch':
#>   query: cell 
#>   requested: 20 (out of 205456)
#>   response(s): 0 

OlsSearch(q = "cell", exact = TRUE)

#> Object of class 'OlsSearch':
#>   query: cell 
#>   requested: 20 (out of 7209)
#>   response(s): 0 

OlsSearch(q = "cell", exact = TRUE, ontology = "go")

#> Object of class 'OlsSearch':
#>   ontolgy: go 
#>   query: cell 
#>   requested: 20 (out of 0)
#>   response(s): 0 

OlsSearch(q = "cell", exact = TRUE, ontology = "GO")

#> Object of class 'OlsSearch':
#>   ontolgy: GO 
#>   query: cell 
#>   requested: 20 (out of 0)
#>   response(s): 0 

OlsSearch(q = "electrospray", ontology = "MS")

#> Object of class 'OlsSearch':
#>   ontolgy: MS 
#>   query: electrospray 
#>   requested: 20 (out of 8)
#>   response(s): 0 

OlsSearch(q = "ionization", ontology = "MS")

#> Object of class 'OlsSearch':
#>   ontolgy: MS 
#>   query: ionization 
#>   requested: 20 (out of 53)
#>   response(s): 0 

OlsSearch(q = "electrospray ionization", ontology = "MS")

#> Object of class 'OlsSearch':
#>   ontolgy: MS 
#>   query: electrospray ionization 
#>   requested: 20 (out of 56)
#>   response(s): 0 

OlsSearch(q = "electrospray ionization", ontology = "MS", exact=TRUE)

#> Object of class 'OlsSearch':
#>   ontolgy: MS 
#>   query: electrospray ionization 
#>   requested: 20 (out of 1)
#>   response(s): 0 

## Request 5 results instead of 20 (default)
OlsSearch(q = "plasma,membrane", ontology = "go", rows = 5)

#> Object of class 'OlsSearch':
#>   ontolgy: go 
#>   query: plasma,membrane 
#>   requested: 5 (out of 4118)
#>   response(s): 0 

## or, once the object was created
(res <- OlsSearch(q = "plasma,membrane", ontology = "go"))

#> Object of class 'OlsSearch':
#>   ontolgy: go 
#>   query: plasma,membrane 
#>   requested: 20 (out of 4118)
#>   response(s): 0 

olsRows(res) <- 5
res

#> Object of class 'OlsSearch':
#>   ontolgy: go 
#>   query: plasma,membrane 
#>   requested: 5 (out of 4118)
#>   response(s): 0 

## all results
res <- allRows(res)
res

#> Object of class 'OlsSearch':
#>   ontolgy: go 
#>   query: plasma,membrane 
#>   requested: 4118 (out of 4118)
#>   response(s): 0 

res <- OlsSearch(q = "trans-golgi", ontology = "go", rows = 5)
res

#> Object of class 'OlsSearch':
#>   ontolgy: go 
#>   query: trans-golgi 
#>   requested: 5 (out of 645)
#>   response(s): 0 

res <- olsSearch(res)
res

#> Object of class 'OlsSearch':
#>   ontolgy: go 
#>   query: trans-golgi 
#>   requested: 5 (out of 645)
#>   response(s): 5 

as(res, "data.frame")

#>                                                   id
#> 1 go:class:http://purl.obolibrary.org/obo/GO_1990676
#> 2 go:class:http://purl.obolibrary.org/obo/GO_0000138
#> 3 go:class:http://purl.obolibrary.org/obo/GO_0005802
#> 4 go:class:http://purl.obolibrary.org/obo/GO_0032588
#> 5 go:class:http://purl.obolibrary.org/obo/GO_0030140
#>                                         iri short_form     obo_id
#> 1 http://purl.obolibrary.org/obo/GO_1990676 GO_1990676 GO:1990676
#> 2 http://purl.obolibrary.org/obo/GO_0000138 GO_0000138 GO:0000138
#> 3 http://purl.obolibrary.org/obo/GO_0005802 GO_0005802 GO:0005802
#> 4 http://purl.obolibrary.org/obo/GO_0032588 GO_0032588 GO:0032588
#> 5 http://purl.obolibrary.org/obo/GO_0030140 GO_0030140 GO:0030140
#>                                   label
#> 1         Golgi trans cisterna membrane
#> 2                  Golgi trans cisterna
#> 3                   trans-Golgi network
#> 4          trans-Golgi network membrane
#> 5 trans-Golgi network transport vesicle
#>                                                                                                                                                                                                                                                                                                                                                                                   description
#> 1                                                                                                                                                                                                                                                                     The lipid bilayer surrounding any of the thin, flattened compartments that form the trans portion of the Golgi complex.
#> 2                                                                                                                                                          The Golgi cisterna farthest from the endoplasmic reticulum; the final processing compartment through which proteins pass before exiting the Golgi apparatus; the compartment in which N-linked protein glycosylation is completed.
#> 3 The network of interconnected tubular and cisternal structures located within the Golgi apparatus on the side distal to the endoplasmic reticulum, from which secretory vesicles emerge. The trans-Golgi network is important in the later stages of protein secretion where it is thought to play a key role in the sorting and targeting of secreted proteins to the correct destination.
#> 4                                                                                                                                                                                                                                                                                                 The lipid bilayer surrounding any of the compartments that make up the trans-Golgi network.
#> 5                                                                                                                                                                                                                                                                                              A vesicle that mediates transport between the trans-Golgi network and other parts of the cell.
#>   ontology_name ontology_prefix  type is_defining_ontology
#> 1            go              GO class                 TRUE
#> 2            go              GO class                 TRUE
#> 3            go              GO class                 TRUE
#> 4            go              GO class                 TRUE
#> 5            go              GO class                 TRUE

res <- as(res, "Terms")
res

#> Object of class 'Terms' with 5 entries
#>  From the GO ontology
#>   GO:1990676, GO:0000138 ... GO:0032588, GO:0030140 

termPrefix(res)

#> GO:1990676 GO:0000138 GO:0005802 GO:0032588 GO:0030140 
#>       "GO"       "GO"       "GO"       "GO"       "GO" 

termId(res)

#>   GO:1990676   GO:0000138   GO:0005802   GO:0032588   GO:0030140 
#> "GO:1990676" "GO:0000138" "GO:0005802" "GO:0032588" "GO:0030140"