Class "Properties"

Propteries (relationships) between terms can be queries for complete Ontology objects and Term/Terms instances, and the results are stored as objects of class Property or Properties.

Objects from the Class

Objects can be created by calls to properties, as described below.

Slots

See the Term and Terms classes.

Extends

Class "Terms", directly.

Methods and functions

properties

signature(object = "Ontology", pagesize = 200): ... Also works with a character with the ontology namespace. See Ontology for details.

properties

signature(object = "Term"): retrieves the properties of term object and returns a Properties object. Returns NULL when no properties are available.

proterties

signature(object = "Terms", ...): retrieves the properties of each term of object and returns a list of Properties (or NULL) items.

show

signature(object = "Properties"): shows a textual summary of the object.

length

signature(object = "Properties"): returns the number of properties in object.

Author

Laurent Gatto <lg390@cam.ac.uk>

Examples

trm <- term("uberon", "UBERON:0002107")
trm
#> A Term from the UBERON ontology: UBERON:0002107 
#>  Label: liver
#>   No description
properties(trm)
#> Object of class 'Properties' with 147 entries
#>  From the UBERON ontology
#>   digestive system gland, abdomen element ... hepatobiliary system, bile 

trm2 <- term("GO", "GO:0005326")
trm2
#> A Term from the GO ontology: GO:0005326 
#>  Label: neurotransmitter transmembrane transporter activity
#>   No description
properties(trm2)
#> Object of class 'Properties' with 18 entries
#>  From the GO ontology
#>   transmembrane transporter activity, neurotransmitter transport ... acetylcholine:proton antiporter activity, neurotransmitter:sodium symporter activity