Class

The Term class describes an ontology term. A set of terms are instantiated as a Terms class.

Objects from the Class

Objects can be created using the term and terms functions. The latter is used with an object of class Ontology or a character describing a valid ontology prefix to download and instantiate all terms of an ontology of interest. The former takes an Ontology object (or an ontology prefix) and a term identifer to instantiate that specific term. See also the 'Methods and functions' sections.

For any given Term object, the children, parents, ancestors, descendants, partOf and derivesFrom terms can be generated and returned as Terms objects.

Terms instances can be subset with [ and [[ and iterated over with lapply.

Slots

iri:

Object of class "character" ~~

label:

Object of class "character" ~~

description:

Object of class "NullOrList" ~~

annotation:

Object of class "list" ~~

synonym:

Object of class "NullOrList" ~~

ontology_name:

Object of class "character" ~~

ontology_prefix:

Object of class "character" ~~

ontology_iri:

Object of class "character" ~~

is_obsolete:

Object of class "logical" ~~

is_defining_ontology:

Object of class "logical" ~~

has_children:

Object of class "logical" ~~

is_root:

Object of class "logical" ~~

short_form:

Object of class "character" ~~

obo_id:

Object of class "NullOrChar" ~~

links:

Object of class "list" ~~

Methods and functions

term

signature(object = "Ontology", id = "character"): ...

terms

signature(x = "Ontology", pagesize = "numeric"): ...

termDesc

signature(object = "Term"): ...

termLabel

signature(object = "Term"): ...

termPrefix

signature(object = "Term"): ...

termSynonym

signature(object = "Term"): ...

termNamespace

signature(object = "Term"): ...

termOntology

signature(object = "Term"): ...

isRoot

signature(object = "Term"): ...

isObsolete

signature(object = "Term"): ...

termId

signature(object = "Term"): ...

children

signature(object = "Term"): Returns a new Terms instance with the object's children. NULL if there are not children.

parents

signature(object = "Term"): Returns a new Terms instance with the object's parents. NULL if there are not parents.

ancestors

signature(object = "Term"): Returns a new Terms instance with the object's ancestors. NULL if there are not ancestors.

descendants

signature(object = "Term"): Returns a new Terms instance with the object's descendants. NULL if there are not descendants.

partOf

signature(object = "Term"): Returns a new Terms instance with terms the object's is a part of. NULL if none.

derivesFrom

signature(object = "Term"): Returns a new Terms instance with terms the object's is derived from. NULL if none.

show

signature(object = "Term"): ...

show

signature(object = "Terms"): ...

all.equal

signature(target = "Term", current = "Term"): ...

all.equal

signature(target = "Terms", current = "Terms"): ...

length

signature(object = "Terms"): returns the number of ontolgies described by the Terms object.

unique

signature(x = "Terms"): returns a new Terms object where all duplicated Term instances, i.e. those with the same term id (even when stemming from different ontologies), have been removed (only the first occurence is retained).

Examples

## (all) terms
(gotrms <- terms("go", pagesize = 10000))

#> Object of class 'Terms' with 47893 entries
#>  From the GO ontology
#>   GO:0005230, GO:0015276 ... GO:0032942, GO:0019987 

## Not run: ------------------------------------
#     ## or, using on ontology object
#     go <- Ontology("go")
#     gotrms <- terms(go, pagesize = 10000)
## ---------------------------------------------

## (one) term
(trm <- gotrms[[1]])

#> A Term from the GO ontology: GO:0005230 
#>  Label: extracellular ligand-gated ion channel activity
#>   Enables the transmembrane transfer of an ion by a channel that opens
#>   when a specific extracellular ligand has been bound by the channel
#>   complex or one of its constituent parts.

termPrefix(trm)

#> [1] "GO"

gotrms[1:3]

#> Object of class 'Terms' with 3 entries
#>  From the GO ontology
#> GO:0005230, GO:0015276, GO:0005231 

gotrms[["GO:0032801"]]

#> A Term from the GO ontology: GO:0032801 
#>  Label: receptor catabolic process
#>   The chemical reactions and pathways resulting in the breakdown of a
#>   receptor molecule, a macromolecule that undergoes combination with a
#>   hormone, neurotransmitter, drug or intracellular messenger to initiate
#>   a change in cell function.

## using an Ontology object
go <- Ontology("GO")
term(go, "GO:0032801")

#> A Term from the GO ontology: GO:0032801 
#>  Label: receptor catabolic process
#>   The chemical reactions and pathways resulting in the breakdown of a
#>   receptor molecule, a macromolecule that undergoes combination with a
#>   hormone, neurotransmitter, drug or intracellular messenger to initiate
#>   a change in cell function.

## using an ontology prefix
term("GO", "GO:0032801")

#> A Term from the GO ontology: GO:0032801 
#>  Label: receptor catabolic process
#>   The chemical reactions and pathways resulting in the breakdown of a
#>   receptor molecule, a macromolecule that undergoes combination with a
#>   hormone, neurotransmitter, drug or intracellular messenger to initiate
#>   a change in cell function.

isObsolete(gotrms[["GO:0030533"]])

#> [1] FALSE

isObsolete(gotrms[["GO:0005563"]])

#> [1] TRUE

isRoot(gotrms[["GO:0030533"]])

#> [1] FALSE

i <- isRoot(gotrms) & !isObsolete(gotrms)
gotrms[i]

#> Object of class 'Terms' with 4 entries
#>  From the GO ontology
#> NULL, GO:0008150, GO:0003674, GO:0005575 

for (ii in which(i))
    show(gotrms[[ii]])

#> A Term from the GO ontology: 
#>  Label: Thing
#>   No description
#> A Term from the GO ontology: GO:0008150 
#>  Label: biological_process
#>   Any process specifically pertinent to the functioning of integrated
#>   living units: cells, tissues, organs, and organisms. A process is a
#>   collection of molecular events with a defined beginning and end.
#> A Term from the GO ontology: GO:0003674 
#>  Label: molecular_function
#>   Elemental activities, such as catalysis or binding, describing the
#>   actions of a gene product at the molecular level. A given gene product
#>   may exhibit one or more molecular functions.
#> A Term from the GO ontology: GO:0005575 
#>  Label: cellular_component
#>   The part of a cell, extracellular environment or virus in which a gene
#>   product is located. A gene product may be located in one or more parts
#>   of a cell and its location may be as specific as a particular
#>   macromolecular complex, that is, a stable, persistent association of
#>   macromolecules that function together.

## or, directly querying the ontology
olsRoot(go)

#> Object of class 'Terms' with 3 entries
#>  From the GO ontology
#> GO:0005575, GO:0003674, GO:0008150 

parents(trm)

#> Object of class 'Terms' with 1 entries
#>  From the GO ontology
#> GO:0015276 

ancestors(trm)

#> Object of class 'Terms' with 20 entries
#>  From the GO ontology
#>   GO:0015276, GO:0022834 ... GO:0005215, GO:0003674 

children(trm)

#> Object of class 'Terms' with 5 entries
#>  From the GO ontology
#>   GO:0036081, GO:0036082 ... GO:0005237, GO:0005231 

descendants(trm)

#> Object of class 'Terms' with 20 entries
#>  From the GO ontology
#>   GO:0005231, GO:0005232 ... GO:0004971, GO:0004972 

partOf(gotrms[["GO:0008308"]])

#> Object of class 'Terms' with 1 entries
#>  From the GO ontology
#> GO:1903959 

partOf(term("BTO", "BTO:0000142"))

#> Object of class 'Terms' with 2 entries
#>  From the BTO ontology
#> BTO:0000227, BTO:0000282 

derivesFrom(term("BTO", "BTO:0002600"))

#> Object of class 'Terms' with 1 entries
#>  From the BTO ontology
#> BTO:0000099 

derivesFrom(term("BTO", "BTO:0001023"))

#> Object of class 'Terms' with 1 entries
#>  From the BTO ontology
#> BTO:0000975