Class "Term"

The Term class describes an ontology term. A set of terms are instantiated as a Terms class.

Objects from the Class

Objects can be created using the term and terms functions. The latter is used with an object of class Ontology or a character describing a valid ontology prefix to download and instantiate all terms of an ontology of interest. The former takes an Ontology object (or an ontology prefix) and a term identifer to instantiate that specific term. See also the 'Methods and functions' sections.

For any given Term object, the children, parents, ancestors, descendants, partOf and derivesFrom terms can be generated and returned as Terms objects.

Terms instances can be subset with [ and [[ and iterated over with lapply.

Slots

iri:

Object of class "character" ~~

label:

Object of class "character" ~~

description:

Object of class "NullOrList" ~~

annotation:

Object of class "list" ~~

synonym:

Object of class "NullOrList" ~~

ontology_name:

Object of class "character" ~~

ontology_prefix:

Object of class "character" ~~

ontology_iri:

Object of class "character" ~~

is_obsolete:

Object of class "logical" ~~

is_defining_ontology:

Object of class "logical" ~~

has_children:

Object of class "logical" ~~

is_root:

Object of class "logical" ~~

short_form:

Object of class "character" ~~

obo_id:

Object of class "NullOrChar" ~~

links:

Object of class "list" ~~

Methods and functions

term

signature(object = "Ontology", id = "character"): ...

terms

signature(x = "Ontology", pagesize = "numeric"): ...

termDesc

signature(object = "Term"): ...

termLabel

signature(object = "Term"): ...

termPrefix

signature(object = "Term"): ...

termSynonym

signature(object = "Term"): ...

termNamespace

signature(object = "Term"): ...

termOntology

signature(object = "Term"): ...

isRoot

signature(object = "Term"): ...

isObsolete

signature(object = "Term"): ...

termId

signature(object = "Term"): ...

children

signature(object = "Term"): Returns a new Terms instance with the object's children. NULL if there are not children.

parents

signature(object = "Term"): Returns a new Terms instance with the object's parents. NULL if there are not parents.

ancestors

signature(object = "Term"): Returns a new Terms instance with the object's ancestors. NULL if there are not ancestors.

descendants

signature(object = "Term"): Returns a new Terms instance with the object's descendants. NULL if there are not descendants.

partOf

signature(object = "Term"): Returns a new Terms instance with terms the object's is a part of. NULL if none.

derivesFrom

signature(object = "Term"): Returns a new Terms instance with terms the object's is derived from. NULL if none.

show

signature(object = "Term"): ...

show

signature(object = "Terms"): ...

all.equal

signature(target = "Term", current = "Term"): ...

all.equal

signature(target = "Terms", current = "Terms"): ...

length

signature(object = "Terms"): returns the number of ontolgies described by the Terms object.

unique

signature(x = "Terms"): returns a new Terms object where all duplicated Term instances, i.e. those with the same term id (even when stemming from different ontologies), have been removed (only the first occurence is retained).

as(x, "data.fram")

Coerces a single Term or Terms into a data.frame of length 1 (for the former) or length length(x) for the latter. The result will contain the following columns: id, label, description of the term(s), their ontology, whether they are obsolete, have children or are root node, the first synonym only, their iri and whether they are defining the ontology. Any missing value will be reported as NA.

Examples

## (all) terms
(gotrms <- terms("go", pagesize = 10000))

#> Object of class 'Terms' with 49161 entries
#>  From the GO ontology
#>   GO:0005230, GO:0015276 ... GO:0032957, GO:0032942 

not_run({
      ## or, using on ontology object
      go <- Ontology("go")
      gotrms <- terms(go, pagesize = 10000)
})

as(gotrms[1:10], "data.frame") ## data,frame with 10 rows

#>                    id
#> GO:0005230 GO:0005230
#> GO:0015276 GO:0015276
#> GO:0005231 GO:0005231
#> GO:0004888 GO:0004888
#> GO:0060079 GO:0060079
#> GO:0005232 GO:0005232
#> NULL             <NA>
#> GO:0005234 GO:0005234
#> GO:0044832 GO:0044832
#> GO:0001819 GO:0001819
#>                                                                 label
#> GO:0005230            extracellular ligand-gated ion channel activity
#> GO:0015276                          ligand-gated ion channel activity
#> GO:0005231 excitatory extracellular ligand-gated ion channel activity
#> GO:0004888                  transmembrane signaling receptor activity
#> GO:0060079                          excitatory postsynaptic potential
#> GO:0005232                                                 GO_0005232
#> NULL                                                            Thing
#> GO:0005234         extracellular-glutamate-gated ion channel activity
#> GO:0044832   positive regulation by virus of host cytokine production
#> GO:0001819                 positive regulation of cytokine production
#>                                                                                                                                                                                                                                                                                                description
#> GO:0005230                                                                                                                Enables the transmembrane transfer of an ion by a channel that opens when a specific extracellular ligand has been bound by the channel complex or one of its constituent parts.
#> GO:0015276                                                                                                                              Enables the transmembrane transfer of an ion by a channel that opens when a specific ligand has been bound by the channel complex or one of its constituent parts.
#> GO:0005231                                        Enables the transmembrane transfer of an ion by a channel that opens when a specific extracellular ligand has been bound by the channel complex or one of its constituent parts, where channel opening contributes to an increase in membrane potential.
#> GO:0004888                                                                                   Combining with an extracellular or intracellular signal and transmitting the signal from one side of the membrane to the other to initiate a change in cell activity or state as part of signal transduction.
#> GO:0060079 A process that leads to a temporary increase in postsynaptic potential due to the flow of positively charged ions into the postsynaptic cell. The flow of ions that causes an EPSP is an excitatory postsynaptic current (EPSC) and makes it easier for the neuron to fire an action potential.
#> GO:0005232                                                                                                                                                                                                                                                                                            <NA>
#> NULL                                                                                                                                                                                                                                                                                                  <NA>
#> GO:0005234                                                                                                                        Enables the transmembrane transfer of an ion by a channel that opens when extracellular glutamate has been bound by the channel complex or one of its constituent parts.
#> GO:0044832                                                                                                                                                                               The process in which a virus increases the frequency, rate or extent of cytokine production in its host organism.
#> GO:0001819                                                                                                                                                                                              Any process that activates or increases the frequency, rate or extent of production of a cytokine.
#>            ontology is_obsolete has_children is_root
#> GO:0005230       go       FALSE         TRUE   FALSE
#> GO:0015276       go       FALSE         TRUE   FALSE
#> GO:0005231       go       FALSE         TRUE   FALSE
#> GO:0004888       go       FALSE         TRUE   FALSE
#> GO:0060079       go       FALSE         TRUE   FALSE
#> GO:0005232       go        TRUE        FALSE    TRUE
#> NULL             go       FALSE        FALSE    TRUE
#> GO:0005234       go       FALSE         TRUE   FALSE
#> GO:0044832       go       FALSE        FALSE   FALSE
#> GO:0001819       go       FALSE         TRUE   FALSE
#>                                                        first_synonym
#> GO:0005230                                                      <NA>
#> GO:0015276                                                      <NA>
#> GO:0005231                                                      <NA>
#> GO:0004888                transmembrane signalling receptor activity
#> GO:0060079 regulation of excitatory post-synaptic membrane potential
#> GO:0005232                                                      <NA>
#> NULL                                                            <NA>
#> GO:0005234                                                      <NA>
#> GO:0044832                                                      <NA>
#> GO:0001819                      up-regulation of cytokine production
#>                                                  iri is_defining_ontology
#> GO:0005230 http://purl.obolibrary.org/obo/GO_0005230                 TRUE
#> GO:0015276 http://purl.obolibrary.org/obo/GO_0015276                 TRUE
#> GO:0005231 http://purl.obolibrary.org/obo/GO_0005231                 TRUE
#> GO:0004888 http://purl.obolibrary.org/obo/GO_0004888                 TRUE
#> GO:0060079 http://purl.obolibrary.org/obo/GO_0060079                 TRUE
#> GO:0005232 http://purl.obolibrary.org/obo/GO_0005232                 TRUE
#> NULL             http://www.w3.org/2002/07/owl#Thing                FALSE
#> GO:0005234 http://purl.obolibrary.org/obo/GO_0005234                 TRUE
#> GO:0044832 http://purl.obolibrary.org/obo/GO_0044832                 TRUE
#> GO:0001819 http://purl.obolibrary.org/obo/GO_0001819                 TRUE

## (one) term
(trm <- gotrms[[1]])

#> A Term from the GO ontology: GO:0005230 
#>  Label: extracellular ligand-gated ion channel activity
#>   Enables the transmembrane transfer of an ion by a channel that opens
#>   when a specific extracellular ligand has been bound by the channel
#>   complex or one of its constituent parts.

termPrefix(trm)

#> [1] "GO"

gotrms[1:3]

#> Object of class 'Terms' with 3 entries
#>  From the GO ontology
#> GO:0005230, GO:0015276, GO:0005231 

gotrms[["GO:0032801"]]

#> A Term from the GO ontology: GO:0032801 
#>  Label: receptor catabolic process
#>   The chemical reactions and pathways resulting in the breakdown of a
#>   receptor molecule, a macromolecule that undergoes combination with a
#>   hormone, neurotransmitter, drug or intracellular messenger to initiate
#>   a change in cell function.

as(trm, "data.frame") ## data,frame with 1 row

#>           id                                           label
#> 1 GO:0005230 extracellular ligand-gated ion channel activity
#>                                                                                                                                                                        description
#> 1 Enables the transmembrane transfer of an ion by a channel that opens when a specific extracellular ligand has been bound by the channel complex or one of its constituent parts.
#>   ontology is_obsolete has_children is_root first_synonym
#> 1       go       FALSE         TRUE   FALSE            NA
#>                                         iri is_defining_ontology
#> 1 http://purl.obolibrary.org/obo/GO_0005230                 TRUE

## using an Ontology object
go <- Ontology("GO")

#> Error in Ontology("GO"): could not find function "Ontology"

term(go, "GO:0032801")

#> Error in term(go, "GO:0032801"): object 'go' not found

## using an ontology prefix
term("GO", "GO:0032801")

#> A Term from the GO ontology: GO:0032801 
#>  Label: receptor catabolic process
#>   The chemical reactions and pathways resulting in the breakdown of a
#>   receptor molecule, a macromolecule that undergoes combination with a
#>   hormone, neurotransmitter, drug or intracellular messenger to initiate
#>   a change in cell function.

isObsolete(gotrms[["GO:0030533"]])

#> [1] FALSE

isObsolete(gotrms[["GO:0005563"]])

#> [1] TRUE

isRoot(gotrms[["GO:0030533"]])

#> [1] FALSE

i <- isRoot(gotrms) & !isObsolete(gotrms)
gotrms[i]

#> Object of class 'Terms' with 4 entries
#>  From the GO ontology
#> NULL, GO:0008150, GO:0003674, GO:0005575 

for (ii in which(i))
    show(gotrms[[ii]])

#> A Term from the GO ontology: 
#>  Label: Thing
#>   No description
#> A Term from the GO ontology: GO:0008150 
#>  Label: biological_process
#>   Any process specifically pertinent to the functioning of integrated
#>   living units: cells, tissues, organs, and organisms. A process is a
#>   collection of molecular events with a defined beginning and end.
#> A Term from the GO ontology: GO:0003674 
#>  Label: molecular_function
#>   The actions of a single gene product or complex at the molecular level
#>   consisting of a single biochemical activity or multiple causally linked
#>   biochemical activities. A given gene product may exhibit one or more
#>   molecular functions.
#> A Term from the GO ontology: GO:0005575 
#>  Label: cellular_component
#>   The part of a cell, extracellular environment or virus in which a gene
#>   product is located. A gene product may be located in one or more parts
#>   of a cell and its location may be as specific as a particular
#>   macromolecular complex, that is, a stable, persistent association of
#>   macromolecules that function together.

## or, directly querying the ontology
olsRoot(go)

#> Error in olsRoot(go): object 'go' not found

parents(trm)

#> Object of class 'Terms' with 1 entries
#>  From the GO ontology
#> GO:0015276 

ancestors(trm)

#> Object of class 'Terms' with 1 entries
#>  From the GO ontology
#> GO:0015276 

children(trm)

#> Object of class 'Terms' with 5 entries
#>  From the GO ontology
#>   GO:0036081, GO:0036082 ... GO:0005237, GO:0005231 

descendants(trm)

#> Object of class 'Terms' with 20 entries
#>  From the GO ontology
#>   GO:0005231, GO:0005234 ... GO:0036082, GO:0036081 

partOf(gotrms[["GO:0008308"]])

#> Object of class 'Terms' with 1 entries
#>  From the GO ontology
#> GO:1903959 

partOf(term("BTO", "BTO:0000142"))

#> Object of class 'Terms' with 2 entries
#>  From the BTO ontology
#> BTO:0000227, BTO:0000282 

derivesFrom(term("BTO", "BTO:0002600"))

#> Object of class 'Terms' with 1 entries
#>  From the BTO ontology
#> BTO:0000099 

derivesFrom(term("BTO", "BTO:0001023"))

#> Object of class 'Terms' with 1 entries
#>  From the BTO ontology
#> BTO:0000975